Move registration tests to public API test file
