+++ code/edgebridge/src/test/java/com/adobe/marketing/mobile/edge/bridge/EdgeBridgeExtensionTests.java
@@ -22,10 +22,7 @@ import com.adobe.marketing.mobile.Event;
 import com.adobe.marketing.mobile.EventSource;
 import com.adobe.marketing.mobile.EventType;
 import com.adobe.marketing.mobile.ExtensionApi;
-import com.adobe.marketing.mobile.ExtensionError;
-import com.adobe.marketing.mobile.ExtensionErrorCallback;
 import com.adobe.marketing.mobile.ExtensionEventListener;
-import com.adobe.marketing.mobile.MobileCore;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -40,7 +37,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
-import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
@@ -64,48 +60,6 @@ public class EdgeBridgeExtensionTests {
 	// ========================================================================================
 	// constructor
 	// ========================================================================================
-	@Test
-	public void test_registerExtension_deprecated() {
-		try (MockedStatic<MobileCore> mobileCoreMockedStatic = Mockito.mockStatic(MobileCore.class)) {
-			// Mock MobileCore.registerExtension()
-			ArgumentCaptor<Class> extensionClassCaptor = ArgumentCaptor.forClass(Class.class);
-			ArgumentCaptor<ExtensionErrorCallback> callbackCaptor = ArgumentCaptor.forClass(
-				ExtensionErrorCallback.class
-			);
-			mobileCoreMockedStatic
-				.when(() -> MobileCore.registerExtension(extensionClassCaptor.capture(), callbackCaptor.capture()))
-				.thenReturn(true);
-			// Call deprecated registerExtension() API
-			EdgeBridge.registerExtension();
-			// Verify: happy path - extension registered is EdgeBridgeExtension
-			assertNotNull(callbackCaptor.getValue());
-			assertEquals(EdgeBridgeExtension.class, extensionClassCaptor.getValue());
-			// Verify: captured error callback accepts valid ExtensionError instance
-			callbackCaptor.getValue().error(ExtensionError.UNEXPECTED_ERROR);
-		}
-	}
-
-	@Test
-	public void test_registerExtension_nullError_deprecated() {
-		try (MockedStatic<MobileCore> mobileCoreMockedStatic = Mockito.mockStatic(MobileCore.class)) {
-			// Mock MobileCore.registerExtension()
-			ArgumentCaptor<Class> extensionClassCaptor = ArgumentCaptor.forClass(Class.class);
-			ArgumentCaptor<ExtensionErrorCallback> callbackCaptor = ArgumentCaptor.forClass(
-				ExtensionErrorCallback.class
-			);
-			mobileCoreMockedStatic
-				.when(() -> MobileCore.registerExtension(extensionClassCaptor.capture(), callbackCaptor.capture()))
-				.thenReturn(true);
-			// Call deprecated registerExtension() API
-			EdgeBridge.registerExtension();
-			// Verify: happy path - extension registered is EdgeBridgeExtension
-			assertNotNull(callbackCaptor.getValue());
-			assertEquals(EdgeBridgeExtension.class, extensionClassCaptor.getValue());
-			// Verify: captured error callback does not throw NPE when passed null ExtensionError
-			callbackCaptor.getValue().error(null);
-		}
-	}
-
 	@Test
 	public void test_listenerRegistration() {
 		extension.onRegistered();
