+++ code/edgebridge/src/test/java/com/adobe/marketing/mobile/edge/bridge/EdgeBridgeTest.java
@@ -20,6 +20,7 @@ import static org.mockito.Mockito.times;
 import android.app.Application;
 import android.content.Context;
 import com.adobe.marketing.mobile.Extension;
+import com.adobe.marketing.mobile.ExtensionError;
 import com.adobe.marketing.mobile.ExtensionErrorCallback;
 import com.adobe.marketing.mobile.MobileCore;
 import java.io.FileInputStream;
@@ -29,6 +30,7 @@ import java.util.Properties;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
@@ -50,7 +52,7 @@ public class EdgeBridgeTest {
 	}
 
 	@Test
-	public void testRegisterExtension_deprecated() {
+	public void test_registerExtension_registersOnce_deprecated() {
 		try (MockedStatic<MobileCore> mobileCoreMockedStatic = Mockito.mockStatic(MobileCore.class)) {
 			EdgeBridge.registerExtension();
 			mobileCoreMockedStatic.verify(
@@ -58,6 +60,48 @@ public class EdgeBridgeTest {
 					MobileCore.registerExtension((Class<? extends Extension>) any(), any(ExtensionErrorCallback.class)),
 				times(1)
 			);
+		}
+	}
+
+	@Test
+	public void test_registerExtension_deprecated() {
+		try (MockedStatic<MobileCore> mobileCoreMockedStatic = Mockito.mockStatic(MobileCore.class)) {
+			// Mock MobileCore.registerExtension()
+			ArgumentCaptor<Class> extensionClassCaptor = ArgumentCaptor.forClass(Class.class);
+			ArgumentCaptor<ExtensionErrorCallback> callbackCaptor = ArgumentCaptor.forClass(
+				ExtensionErrorCallback.class
+			);
+			mobileCoreMockedStatic
+				.when(() -> MobileCore.registerExtension(extensionClassCaptor.capture(), callbackCaptor.capture()))
+				.thenReturn(true);
+			// Call deprecated registerExtension() API
+			EdgeBridge.registerExtension();
+			// Verify: happy path - extension registered is EdgeBridgeExtension
+			assertNotNull(callbackCaptor.getValue());
+			assertEquals(EdgeBridgeExtension.class, extensionClassCaptor.getValue());
+			// Verify: captured error callback accepts valid ExtensionError instance
+			callbackCaptor.getValue().error(ExtensionError.UNEXPECTED_ERROR);
+		}
+	}
+
+	@Test
+	public void test_registerExtension_nullError_deprecated() {
+		try (MockedStatic<MobileCore> mobileCoreMockedStatic = Mockito.mockStatic(MobileCore.class)) {
+			// Mock MobileCore.registerExtension()
+			ArgumentCaptor<Class> extensionClassCaptor = ArgumentCaptor.forClass(Class.class);
+			ArgumentCaptor<ExtensionErrorCallback> callbackCaptor = ArgumentCaptor.forClass(
+				ExtensionErrorCallback.class
+			);
+			mobileCoreMockedStatic
+				.when(() -> MobileCore.registerExtension(extensionClassCaptor.capture(), callbackCaptor.capture()))
+				.thenReturn(true);
+			// Call deprecated registerExtension() API
+			EdgeBridge.registerExtension();
+			// Verify: happy path - extension registered is EdgeBridgeExtension
+			assertNotNull(callbackCaptor.getValue());
+			assertEquals(EdgeBridgeExtension.class, extensionClassCaptor.getValue());
+			// Verify: captured error callback does not throw NPE when passed null ExtensionError
+			callbackCaptor.getValue().error(null);
 		}
 	}
